requirements.txt
# --- Core backend framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
pillow==10.4.0

# --- AI/ML & APIs ---
cohere==5.9.4
groq==0.12.0
beautifulsoup4==4.12.3
googlesearch-python==1.2.5
mtranslate==1.8

# --- Automation & Utilities ---
pywhatkit==5.4
AppOpener==1.7
keyboard==0.13.5
pygame
edge-tts==6.1.12
selenium==4.23.1
webdriver-manager==4.0.2
rich==13.7.1

# --- GUI (desktop mode) ---
PyQt5==5.15.11
